MicroCeph Remote Replication (2/3): RBD Mirroring (#437)
# Description
Adds implementation for enabling/disabling rbd mirroring (remote
replication) to a configured MicroCeph remote cluster.

## Type of change

Please delete options that are not relevant.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] CleanCode (Code refactor, test updates, does not introduce
functional changes)
- [X] Documentation update (Contains Doc change)

## How Has This Been Tested?
- [X] CI Tests
- [X] Unit Tests

## Contributor's Checklist

Please check that you have:

- [ ] self-reviewed the code in this PR.
- [X] added code comments, particularly in hard-to-understand areas.
- [X] updated the user documentation with corresponding changes.
- [X] added tests to verify effectiveness of this change.

---------

+++ microceph/api/remote.go
@@ -106,8 +106,9 @@ func cmdRemoteDelete(state state.State, r *http.Request) response.Response {
 		return response.BadRequest(err)
 	}
 
-	// Note(utkarshbhatthere): TODO for when remote replication is implemented.
-	// [ ] add check for remote replication pairs before deleting remotes.
+	if isRemoteConfigured(remoteName) {
+		return response.SmartError(fmt.Errorf("cannot remote remote(%s), disable RBD mirroring", remoteName))
+	}
 
 	// Remove remote record.
 	err = database.DeleteRemoteDb(r.Context(), state, remoteName)
@@ -125,6 +126,11 @@ func cmdRemoteDelete(state state.State, r *http.Request) response.Response {
 }
 
 /*****************HELPER FUNCTIONS**************************/
+
+func isRemoteConfigured(remoteName string) bool {
+	// check remote configured for RBD mirroring
+	return ceph.IsRemoteConfiguredForRbdMirror(remoteName)
+}
 
 // renderConfAndKeyringFiles generates the $cluster.conf and $cluster.keyring files on the host.
 func renderConfAndKeyringFiles(remoteName string, localName string, configs map[string]string) error {
